+++ workers/unity/Packages/com.improbable.gdk.gameobjectcreation/LinkedGameObjectMapSubscriptionManager.cs
@@ -1,0 +1,42 @@
+using Improbable.Gdk.Core;
+using Improbable.Gdk.Subscriptions;
+using Unity.Entities;
+
+namespace Improbable.Gdk.GameObjectCreation
+{
+    [AutoRegisterSubscriptionManager]
+    internal class LinkedGameObjectMapSubscriptionManager : SubscriptionManager<LinkedGameObjectMap>
+    {
+        private readonly World world;
+        private Subscription<LinkedGameObjectMap> linkedGameObjectMapSubscription;
+
+        public LinkedGameObjectMapSubscriptionManager(World world)
+        {
+            this.world = world;
+        }
+
+        public override Subscription<LinkedGameObjectMap> Subscribe(EntityId entityId)
+        {
+            if (linkedGameObjectMapSubscription == null)
+            {
+                linkedGameObjectMapSubscription = new Subscription<LinkedGameObjectMap>(this, new EntityId(0));
+
+                var goSystem = world.GetExistingSystem<GameObjectInitializationSystem>();
+                if (goSystem != null)
+                {
+                    linkedGameObjectMapSubscription.SetAvailable(new LinkedGameObjectMap(goSystem.Linker));
+                }
+            }
+
+            return linkedGameObjectMapSubscription;
+        }
+
+        public override void Cancel(ISubscription subscription)
+        {
+        }
+
+        public override void ResetValue(ISubscription subscription)
+        {
+        }
+    }
+}